refactor/own-04 : ApiRes 적용 (#38)
* refactor: ApiRes 적용

* revert: SpringDoc 이름변경 복구
