+++ src/main/java/com/back/orderplz_01/coffee/service/CoffeeService.java
@@ -1,12 +1,12 @@
 package com.back.orderplz_01.coffee.service;
 
 
-import com.back.orderplz_01.coffee.dto.CoffeeDetailResponse;
+import com.back.orderplz_01.coffee.dto.response.CoffeeDetailResponse;
 import com.back.orderplz_01.coffee.entity.Coffee;
 import com.back.orderplz_01.coffee.repository.CoffeeRepository;
 import jakarta.persistence.EntityNotFoundException;
-import com.back.orderplz_01.coffee.dto.CoffeeResponseDto;
-import com.back.orderplz_01.coffee.dto.CoffeeUpdateRequestDto;
+import com.back.orderplz_01.coffee.dto.response.CoffeeResponseDto;
+import com.back.orderplz_01.coffee.dto.request.CoffeeUpdateRequestDto;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
